requirements.txt
blinker==1.6.2
click==8.1.6
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.6
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
typing_extensions==4.7.1
SQLAlchemy==2.0.19
Flask-Migrate==4.0.4
alembic==1.11.1
Mako==1.2.4
bidict-0.22.1 
flask-socketio-5.3.4 
python-engineio-4.5.1 
python-socketio-5.8.0
gevent==23.7.0
gevent-websocket==0.10.1